CAM: move find methods to separate util module
