+++ src/Mod/CAM/CMakeLists.txt
@@ -119,6 +119,11 @@ SET(PathPythonTools_SRCS
     Path/Tool/Controller.py
 )
 
+SET(PathPythonToolsToolBit_SRCS
+    Path/Tool/toolbit/__init__.py
+    Path/Tool/toolbit/util.py
+)
+
 SET(PathPythonToolsGui_SRCS
     Path/Tool/Gui/__init__.py
     Path/Tool/Gui/Bit.py
@@ -438,6 +443,7 @@ SET(all_files
     ${PathPythonTools_SRCS}
     ${PathPythonToolsGui_SRCS}
     ${PathPythonToolsShape_SRCS}
+    ${PathPythonToolsToolBit_SRCS}
     ${PathPythonGui_SRCS}
     ${Tools_SRCS}
     ${Tools_Bit_SRCS}
@@ -578,6 +584,13 @@ INSTALL(
         ${PathPythonToolsShape_SRCS}
     DESTINATION
         Mod/CAM/Path/Tool/Shape
+)
+
+INSTALL(
+    FILES
+        ${PathPythonToolsToolBit_SRCS}
+    DESTINATION
+        Mod/CAM/Path/Tool/toolbit
 )
 
 INSTALL(
